chore(ddl) change id column type to serial

+++ src/main/resources/static/initial_ddl.sql
@@ -3,18 +3,13 @@
 CREATE SCHEMA public AUTHORIZATION pg_database_owner;
 
 COMMENT ON SCHEMA public IS 'standard public schema';
+
 -- public.t_city definition
-
--- Drop table
-
--- DROP TABLE public.t_city;
-
 CREATE TABLE public.t_city (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	"name" varchar(255) NULL,
 	CONSTRAINT city_pkey PRIMARY KEY (id)
 );
-
 
 -- public.t_movie definition
 
@@ -23,7 +18,7 @@ CREATE TABLE public.t_city (
 -- DROP TABLE public.t_movie;
 
 CREATE TABLE public.t_movie (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	title varchar(255) NULL,
 	synopsis varchar(255) NULL,
 	genre _varchar NULL,
@@ -40,7 +35,7 @@ CREATE TABLE public.t_movie (
 -- DROP TABLE public.t_seat_element;
 
 CREATE TABLE public.t_seat_element (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	"name" varchar(255) NULL, -- DOOR / SEAT / STAIRS
 	CONSTRAINT seat_element_pkey PRIMARY KEY (id)
 );
@@ -57,7 +52,7 @@ COMMENT ON COLUMN public.t_seat_element."name" IS 'DOOR / SEAT / STAIRS';
 -- DROP TABLE public.t_seat_type;
 
 CREATE TABLE public.t_seat_type (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	"name" varchar(255) NULL,
 	CONSTRAINT seat_type_pkey PRIMARY KEY (id)
 );
@@ -70,7 +65,7 @@ CREATE TABLE public.t_seat_type (
 -- DROP TABLE public.t_studio_type;
 
 CREATE TABLE public.t_studio_type (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	"name" varchar(255) NULL,
 	row_seats int4 NULL,
 	col_seats int4 NULL,
@@ -85,7 +80,7 @@ CREATE TABLE public.t_studio_type (
 -- DROP TABLE public.t_user;
 
 CREATE TABLE public.t_user (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	"name" varchar(255) NULL,
 	email varchar(255) NULL,
 	mobile_no varchar(255) NULL,
@@ -106,7 +101,7 @@ COMMENT ON COLUMN public.t_user."type" IS 'ADMIN / USER';
 -- DROP TABLE public.t_cinema;
 
 CREATE TABLE public.t_cinema (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	city_id int8 NULL,
 	"name" varchar(255) NULL,
 	CONSTRAINT cinema_pkey PRIMARY KEY (id),
@@ -121,7 +116,7 @@ CREATE TABLE public.t_cinema (
 -- DROP TABLE public.t_studio;
 
 CREATE TABLE public.t_studio (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	cinema_id int8 NULL,
 	type_id int8 NULL,
 	"name" varchar(255) NULL,
@@ -140,7 +135,7 @@ CREATE TABLE public.t_studio (
 -- DROP TABLE public.t_movie_studio_schedule;
 
 CREATE TABLE public.t_movie_studio_schedule (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	studio_id int8 NULL,
 	movie_id int8 NULL,
 	"day" varchar(255) NULL,
@@ -160,7 +155,7 @@ CREATE TABLE public.t_movie_studio_schedule (
 -- DROP TABLE public.t_seat_studio;
 
 CREATE TABLE public.t_seat_studio (
-	id int8 NOT NULL,
+	id serial NOT NULL,
 	seat_type_id int8 NULL,
 	seat_element_id int8 NULL,
 	studio_id int8 NULL,
